backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Google Cloud - Native Stack (muhtar-5ab9b project)
google-cloud-vision==3.5.0
google-cloud-storage==2.14.0
google-cloud-aiplatform==1.42.1  # Vertex AI SDK for Gemini

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Environment & Config
python-dotenv==1.0.1
pydantic==2.6.0
pydantic-settings==2.1.0

# Date/Time
python-dateutil==2.8.2

# Excel Export
openpyxl==3.1.2
pandas==2.2.0

# Image Processing
Pillow==10.2.0

# CORS
starlette==0.35.1
